turn variables switched to ulong
Former-commit-id: ea617e5f4b9987d281173e04dc61145ee669e3ad

+++ Unity Project/Assets/Scripts/MANAGERS/TimeManager.cs
@@ -45,7 +45,7 @@ public class TimeManager : MonoBehaviour, IManager
         }
     }
 
-    public int turnsPassed = 0;
+    public ulong turnsPassed = 0;
     public int numTilesCrossed = 0;
     #endregion
     #region Action Costs
